Move azure.worker to azure.functions_worker
Fixes: #140
