+++ .flake8
@@ -2,4 +2,4 @@
 ignore = W503,E402,E731
 exclude =
     .git, __pycache__, build, dist, .eggs, .github, .local,
-    Samples, azure/worker/protos/, docs/
+    Samples, azure/functions_worker/protos/, docs/
