+++ azure/functions_worker/bindings/__init__.py
