+++ azure/functions_worker/bindings/blob.py
