+++ azure/functions_worker/bindings/context.py
