+++ azure/functions_worker/bindings/cosmosdb.py
