+++ azure/functions_worker/bindings/eventhub.py
