+++ azure/functions_worker/bindings/http.py
