+++ azure/functions_worker/bindings/out.py
