+++ azure/functions_worker/bindings/queue.py
