+++ azure/functions_worker/bindings/timer.py
