+++ azure/functions_worker/protos/.gitignore
