+++ azure/functions_worker/protos/_src/.gitignore
