+++ azure/functions_worker/protos/_src/src/proto/google/protobuf/duration.proto
