+++ setup.cfg
@@ -13,8 +13,8 @@ warn_return_any = True
 disallow_subclassing_any = True
 ignore_missing_imports = True
 
-[mypy-azure.worker.aio_compat]
+[mypy-azure.functions_worker.aio_compat]
 ignore_errors = True
 
-[mypy-azure.worker.protos*]
+[mypy-azure.functions_worker.protos.*]
 ignore_errors = True
