+++ tests/test_broken_functions.py
@@ -1,5 +1,5 @@
-from azure.worker import protos
-from azure.worker import testutils
+from azure.functions_worker import protos
+from azure.functions_worker import testutils
 
 
 class TestMockHost(testutils.AsyncTestCase):
