+++ tests/test_code_quality.py
@@ -16,7 +16,7 @@ class TestCodeQuality(unittest.TestCase):
 
         try:
             subprocess.run(
-                [sys.executable, '-m', 'mypy', '-m', 'azure.worker'],
+                [sys.executable, '-m', 'mypy', '-m', 'azure.functions_worker'],
                 check=True,
                 stdout=subprocess.PIPE,
                 stderr=subprocess.PIPE,
