+++ tests/test_eventhub_functions.py
@@ -1,7 +1,7 @@
 import json
 import time
 
-from azure.worker import testutils
+from azure.functions_worker import testutils
 
 
 class TestEventHubFunctions(testutils.WebHostTestCase):
